Fix formatting for a L2 heading
Something happens consistently across every tool and editor I use when I type a heading as `## The heading` too quickly, it doesn't render as a heading later in markdown. I'm starting to think that OS X sometimes inserts a different space character, but I haven't yet had the opportunity to prove it. Anyway, the solution is to delete the space and type it again. _shrug_